requirements.txt
requests
pytesseract
python-dotenv
pillow
watchdog
openai
